Extract a ContentScript class
